Version fix (#55)
* [IMP] version modules security changes

* [IMP] product_variant_custom_sale: allowed product attribute line
editing in 'sent' state sale lines

* [IMP] product_variant_custom_sale: sale_line_description

* [FIX] product_variant_custom_sale: create button in sale line

* [FIX-IMP] product_version modules: posibility to add product version in new picking + move menu of purchase version + product version onchange fix + AttributeError fix

* [FIX+IMP] many version modules fixes: permissions + sale line descriptions + minor changes. Added new field in pickings to assign product version when picking is created manuali.

+++ product_variant_custom_purchase/models/purchase.py
@@ -79,6 +79,11 @@ class PurchaseOrderLine(models.Model):
     @api.onchange('product_tmpl_id')
     def onchange_product_template(self):
         self.ensure_one()
+        if self.product_tmpl_id and \
+                self.product_id.product_tmpl_id == self.product_tmpl_id:
+            return {'domain': {'product_id':
+                               [('product_tmpl_id', '=',
+                                 self.product_tmpl_id.id)]}}
         self.product_attribute_ids = \
             self._delete_product_attribute_ids()
         self.custom_value_ids = self._delete_custom_lines()
@@ -100,11 +105,16 @@ class PurchaseOrderLine(models.Model):
 
     @api.onchange('product_id')
     def onchange_product_id(self):
+        p1 = self.product_id
         result = super().onchange_product_id()
-        self.custom_value_ids = self._delete_custom_lines()
-        self.product_attribute_ids = self._delete_product_attribute_ids()
+        if not self.product_tmpl_id and self.product_id:
+            self.product_tmpl_id = self.product_id.product_tmpl_id
+            self.product_id = p1
         if self.product_id:
-            product = self.product_id
+            self.custom_value_ids = self._delete_custom_lines()
+            self.product_attribute_ids = self._delete_product_attribute_ids()
+            product = p1
+
             self.product_attribute_ids = \
                 product._get_product_attributes_values_dict()
